src/com/company/bookmark/entities/WebLink.java
package com.company.bookmark.entities;

public class WebLink extends Bookmark{
    private String url;
    private String host;

    public String getUrl() {
        return url;
    }

    public void setUrl(String url) {
        this.url = url;
    }

    @Override
    public String toString() {
        return "WebLink{" +
                "url='" + url + '\'' +
                ", host='" + host + '\'' +
                '}';
    }

    @Override
    public boolean isKidFriendly() {
        return !url.contains("porn") && !getTitle().contains("porn") && !host.contains("adult");
    }

    public String getHost() {
        return host;
    }

    public void setHost(String host) {
        this.host = host;
    }
}
